Infinite input support WIP, updated library status, condensed code
The status of this Library is now working, which means that there are very few bugs. One known bug is with the sub() function and the infinite input support.
Only add() has a fully working infinite input support.
sub() has an "implemented" type status with the infinite input support.
Every other function doesn't have any form of infinite input support, but all except for miniadd() and minisub() will be getting it. Please note that the mini functions are going to be removed as they're only used once each, which means that there's no reason to have them as functions, and aren't ment for other users using the library to have access to it in the first place.