Applying optimizations pt. 2
Optimized:
`multiply`
`formatOutput`
`add`
`subtract`
`shouldRun`

The function `shouldRun` has been overhauled to work much faster. Instead of having a max integer, there's now a max integer length. The default is 15. This is because 16 is the length of the 64-bit limit, but 9999999999999999 is way over the limit, so instead, it's 999999999999999